+++ lib/src/senzu_player.dart
@@ -29,7 +29,7 @@ export 'package:senzu_player/src/data/language/language.dart';
 export 'package:senzu_player/src/controllers/senzu_player_bundle.dart';
 export 'package:senzu_player/src/controllers/senzu_downloader.dart';
 export 'package:senzu_player/src/data/database/download_database.dart'
-    show DownloadTask;
+    show DownloadTask, DownloadImageCache;
 export 'package:senzu_player/src/cast/senzu_cast_controller.dart';
 export 'package:senzu_player/src/cast/senzu_cast_service.dart';
 export 'package:senzu_player/src/data/models/senzu_metadata.dart';